change folder name to schemas
