+++ app/schemas/Accommodation.py
@@ -1,8 +1,4 @@
-from datetime import datetime
-
 from pydantic import BaseModel
-
-from app.domain.Amenitie import Amenitie
 
 
 class AccommodationUpdateDTO(BaseModel):
@@ -23,15 +19,3 @@ class AccommodationCreateDTO(BaseModel):
     double_beds: int
     price: float
     amenities: list[str]
-
-
-class Accommodation(BaseModel):
-    id: int
-    created_at: datetime
-    name: str
-    status: str
-    total_guests: int
-    single_beds: int
-    double_beds: int
-    price: float
-    amenities: list[Amenitie]
