+++ app/schemas/Guest.py
@@ -1,6 +1,3 @@
-from datetime import datetime
-from uuid import UUID
-
 from pydantic import BaseModel
 
 
@@ -13,16 +10,6 @@ class GuestUpdateDTO(BaseModel):
 
 
 class GuestCreateDTO(BaseModel):
-    document: str
-    name: str
-    surname: str
-    phone: str
-    country: str
-
-
-class Guest(BaseModel):
-    uuid: UUID
-    created_at: datetime
     document: str
     name: str
     surname: str
